--- edgy/contrib/admin/app.py
@@ -1,33 +1,0 @@
-from __future__ import annotations
-
-from typing import Any
-
-from lilya.apps import Lilya
-from lilya.requests import Request
-from lilya.routing import Include, RoutePath
-from lilya.templating.controllers import templates  # noqa
-
-from edgy.conf import settings
-from edgy.contrib.admin.views import AdminDashboard, ModelListView
-
-
-async def not_found(request: Request, exc: Exception) -> Any:
-    return templates.get_template_response(
-        request,
-        "404.html",
-        context={"title": "Not Found"},
-    )
-
-
-app = Lilya(
-    routes=[
-        Include(
-            path=settings.admin_config.admin_prefix_url,
-            routes=[
-                RoutePath("/", handler=AdminDashboard, name="admin"),
-                RoutePath("/models", handler=ModelListView, name="models"),
-            ],
-        ),
-    ],
-    exception_handlers={404: not_found},
-)
